gomu::Application: Do not set arbitrary clear color, do not automatically call glClear() before draw.
The arbitrary clear color was red. It's not a good default.
If we don't call glClearColor(), the default is black.

Do not automatically call glClear() before draw, because the user might not
want to clear the screen for various reasons. It also doesn't need to be
called if the whole screen is filled with drawings. Let the user decide.